Deduplicate tool permission logic between agent and agent_servers
Move shared tool permission decision logic into agent_settings so both
the agent crate and agent_servers crate can reuse it, eliminating the
duplicated AcpPermissionDecision enum and check_acp_tool_permission
function.

Changes:
- Add ToolPermissionDecision enum, decide_tool_permission(), and helpers
  (check_commands, check_invalid_patterns) to agent_settings
- Simplify tool_permissions.rs to re-export from agent_settings and keep
  only thin wrappers (decide_permission_from_settings, decide_permission_for_path)
- Remove AcpPermissionDecision and check_acp_tool_permission from acp.rs,
  replacing with calls to the shared decide_tool_permission
- Fix pre-existing compile error (is_empty on IntoIterator) using had_commands bool
- Remove unused shell_command_parser dep from agent_servers
- Remove unused agent dep from settings_ui
