+++ crates/settings_ui/Cargo.toml
@@ -16,7 +16,6 @@ default = []
 test-support = []
 
 [dependencies]
-agent.workspace = true
 agent_settings.workspace = true
 anyhow.workspace = true
 bm25 = "2.3.2"
